move things to DTO namespace
