+++ lib/DTO/NodeConfiguration.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Einenlum\PhpStackDetector\DTO;
+
+use Einenlum\PhpStackDetector\NodePackageManagerType;
+
+readonly class NodeConfiguration
+{
+    public function __construct(
+        public ?string $version,
+        public ?string $requirements,
+        public NodePackageManagerType $packageManager,
+    ) {
+    }
+}
