+++ lib/Detector.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Einenlum\PhpStackDetector;
 
+use Einenlum\PhpStackDetector\DTO\FullConfiguration;
+use Einenlum\PhpStackDetector\DTO\NodeConfiguration;
+use Einenlum\PhpStackDetector\DTO\PhpConfiguration;
 use Einenlum\PhpStackDetector\DTO\Stack;
 
 readonly class Detector
