--- lib/PhpVersion.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Einenlum\PhpStackDetector;
-
-readonly class PhpVersion
-{
-    public function __construct(
-        public ?string $version,
-        public ?string $requirements,
-    ) {
-    }
-}
